+++ R/model_get_nlevels.R
@@ -1,8 +1,7 @@
 #' Get the number of levels for each factor used in `xlevels`
 #'
 #' @param model a model object
-#' @result a tibble with two columns: `"variable"` and `"var_nlevels"`,
-#' `NULL` if no categorical variable
+#' @result a tibble with two columns: `"variable"` and `"var_nlevels"`
 #' @export
 #' @family model_helpers
 #' @examples
@@ -16,7 +15,11 @@ model_get_nlevels <- function(model) {
 #' @rdname model_get_xlevels
 model_get_nlevels.default <- function(model) {
   nlevels <- model_get_xlevels(model) %>% lapply(length)
-  if (length(nlevels) == 0) return(NULL)
+  if (length(nlevels) == 0)
+    return(
+      dplyr::tibble(variable = NA_character_, var_nlevels = NA_integer_) %>%
+        dplyr::filter(FALSE) # empty tibble
+    )
   dplyr::tibble(
     variable = names(nlevels),
     var_nlevels = unlist(nlevels)